Improve dbc lookup
